Rearranged blog layout
