+++ app/javascript/src/App.jsx
@@ -4,8 +4,8 @@ import { QueryClientProvider } from "react-query";
 import { Route, Switch, BrowserRouter as Router } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 
-import { BlogList, CreateBlog, ShowBlog } from "./components/Blog";
-import { PageNotFound } from "./components/common";
+import { Blogs, CreateBlog, ShowBlog } from "./components/Blog";
+import { PageNotFound, Container } from "./components/common";
 import routes from "./route";
 import queryClient from "./utils/queryClient";
 
@@ -13,13 +13,15 @@ const App = () => (
   <QueryClientProvider client={queryClient}>
     <Router>
       <ToastContainer />
-      <Switch>
-        <Route exact path={routes.root} render={() => <h1>Home</h1>} />
-        <Route exact component={BlogList} path={routes.blogs} />
-        <Route exact component={CreateBlog} path={routes.createBlog} />
-        <Route exact component={ShowBlog} path={routes.showBlog} />
-        <Route exact component={PageNotFound} path={routes.anyPath} />
-      </Switch>
+      <Container>
+        <Switch>
+          <Route exact path={routes.root} render={() => <h1>Home</h1>} />
+          <Route exact component={Blogs} path={routes.blogs} />
+          <Route exact component={CreateBlog} path={routes.createBlog} />
+          <Route exact component={ShowBlog} path={routes.showBlog} />
+          <Route exact component={PageNotFound} path={routes.anyPath} />
+        </Switch>
+      </Container>
     </Router>
   </QueryClientProvider>
 );
